Add feature for resampling volumes with gantry tilt
Currently, the resampled output will have the same extent as the input
volume, so part of the original image data can be clipped outside.
The affine transform seems to work, but should be tested on more
datasets.

Non-tilted volumes will just be interpolated, so this feature can
also be used to double the number of slices in an image.

Other changes:
- Add option to save grayscale volumes in VTK format
- Add a CommandManager and limit the maximum undo length to 8
  to avoid running out of memory when working with larger volumes